stable scraper and api w/comments
