+++ client/src/components/Feed.jsx
@@ -2,12 +2,14 @@ import React, { useState } from 'react';
 import { FaBars, FaTimes } from 'react-icons/fa';
 import Logo from '../assets/Gamer Space-logos_transparent.png';
 import { Link } from 'react-router-dom';
-
+import {bot} from '../utils/scraper'
+//returns undefined
+console.log(bot)
 // page for news scrapper
 const Feed = () => {
   const [nav, setNav] = useState(false)
   const handleClick = () => setNav(!nav)
-
+  
   return (
     <div>
       <div className='fixed w-full h-[80px] flex justify-between items-center px-4 text-gray-300 font-serif '>
